app/requirements.txt
#Django==3.2
Django>=4.1.10
gunicorn~=20.1
uwsgi==2.0.20
asgiref==3.6.0
django-split-settings==1.2.0
flake8==4.0.0
mypy==1.2.0
mypy-extensions==1.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.21.1
pytz==2022.7.1
split-settings==1.0.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.5.0
django-extensions~=3.2.0
django-debug-toolbar
bootstrap4
django-bootstrap4
python-dateutil
markdown
Pygments
